Split Language\Util methods into new Database\Util
